norme et clean un peu tout
